+++ 09-the-first-practice-project/react-investment/src/components/Header/Header.js
@@ -1,4 +1,4 @@
-import logo from "../assets/investment-calculator-logo.png";
+import logo from "../../assets/investment-calculator-logo.png";
 
 const Header = () => (
   <header className="header">
